Add a RAM usage scale factor configurable from the command line (#374)
* track block txs by approx bytes

* 0.5 scale factor on byte budgets

* fix obsolete comment (unrelated to this pr)

* reduce orphan block flood

* use tracked bytes mode for acceptance data cache

* impl `estimate_mem_units` only when actually used

* propagate ram scale arg from the command line

* add to simpa as well

* make it clear that mass is an average

* fix cpu usage stat

+++ consensus/src/pipeline/monitor.rs
@@ -59,7 +59,7 @@ impl ConsensusMonitor {
                 if delta.header_counts != 0 { delta.dep_counts as f64 / delta.header_counts as f64 } else { 0f64 },
                 if delta.header_counts != 0 { delta.mergeset_counts as f64 / delta.header_counts as f64 } else { 0f64 },
                 if delta.body_counts != 0 { delta.txs_counts as f64 / delta.body_counts as f64 } else{ 0f64 },
-                if delta.body_counts != 0 { delta.mass_counts / delta.body_counts } else{ 0 },
+                if delta.body_counts != 0 { delta.mass_counts as f64 / delta.body_counts as f64 } else{ 0f64 },
             );
 
             last_snapshot = snapshot;
